# src/modules/tasks/dto/create-task.dto.ts
import { IsNumber, IsOptional, IsString } from "class-validator";

export class CreateTaskDto {
  @IsString()
  title!: string;

  @IsOptional()
  @IsString()
  description!: string;

  @IsNumber()
  userId!: number;
}